+++ extensions/model-context-protocol-registry/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Model Context Protocol Registry Changelog
 
+## [Add Paperless MCP Server] - 2025-05-21
+
+Add Community Paperless MCP Server to registry, this allows to manage documents in Paperless.
+
+## [Add Shopify Dev Server] - 2025-05-21
+
+Add Official Shopify Dev MCP Server to registry.
+
+## [Add Prisma MCP Server] - 2025-05-21
+
+Add Official Prisma MCP Server to registry.
+
+## [Add Zeabur MCP Server] - 2025-05-21
+
+Add Official Zeabur MCP Server to registry, this allows to manage and deploy Zeabur projects.
+
+## [Add Apify MCP Server] - 2025-05-16
+
+Add the official Apify MCP server to the Model Context Protocol Registry. Apify MCP server allows AI agents to use 5,000+ ready-made Actors for use cases such as extracting data from websites, social media, search engines, online maps, and more.
+
+## [Add Nuxt MCP Server] - 2025-05-15
+
+Add Official Nuxt MCP Server to registry, this allow docs and modules search.
+
 ## [Chore: Fixed broken MCP links] - 2025-05-09
 
 ## [Initial Version] - 2025-05-09