+++ extensions/window-sizer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Window Sizer Changelog
 
+## [Fixes] - 2025-05-23
+
+- Fixed occasional issue with action icon not displaying
+
+## [UI improvement] - 2025-05-21
+
+- Adjusted Action icon to only display when selected
+
+## [Starred Sizes & Improvements] - 2025-05-19
+
+- Added starring for frequently used window sizes
+- Unified shortcut keys for Deleting Custom and Starred Sizes
+- Improved action icons
+
 ## [New Action Icons] - 2025-05-12
 
 - Added new action icons